Add navbar with login/logout and home icon
